Relocated downloader pkg
